refactor(editor): add rich text toolbar plugin tests and utility functions
- Introduced `rich-toolbar-plugin-test-utils.tsx` for common test utilities.
- Added tests for inline text formatting in `rich-toolbar-plugin-text-formatting.test.tsx`.
- Removed the old `rich-toolbar-plugin.test.tsx` to streamline test organization.
- Created `toolbar-divider.tsx`, `toolbar-icon-button.tsx`, and `toolbar-select.tsx` components for improved toolbar structure and functionality.
